Allow Users to Login, Create, and Update their accounts with UI
Changes:
Index.html now shows elements to post an account to logged in users,
otherwise it'll only show a login or create account link.

SetUserInfo.html is newly created, it allows users to edit an existing account,
or lets new users create an account on the site.

Purpose:
In the future Comments will show who posted them, so this will let users
create accounts and log in to post comments.

Concerns:
People need to create an account on the setUserInfo page, logging in is insufficent.
Users who have never created an account but "login" will not be recgonized as logged in.